image/dups/requirements.txt
imagededup
python-slugify
